DistroKit: update to ptxdist-2017.05.0
In ptxdist-2017.05.0, the list of packages in DistroKit has been
systematically reviewed, so all packages are now available in their
recent versions.

There are two exceptions:
ncurses 6.0 and networkmanager 1.8 didn't make it into the release.

As networkmanager 1.8 is now available in the BSP, we can make use of
its newly reduced dependencies and switch off libsoup, sqlite and
libxml2.

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>