doc: extend documentation for supported hardware
The documentation does now contain the names of the images and
information about where to find the serial console ports of the boards.

As some hardware platforms do currently not work again after the recent
rework, we remove them from the ToC for now and re-add them later when
they have been tested and fixed.

Adding Build doc to .gitignore.

Signed-off-by: Chris Fiege <c.fiege@pengutronix.de>
[improve commit message]
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>