kernelconfig: reason: enable SCHED_STACK_END_CHECK
This option checks for a stack overrun on calls to schedule().
If the stack end location is found to be over written always panic as
the content of the corrupted region can no longer be trusted.
This is to ensure no erroneous behaviour occurs which could result in
data corruption or a sporadic crash at a later stage once the region
is examined. The runtime overhead introduced is minimal.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>