scripts: backport bugfix for /etc/issue generation
In 4e3f122b9094f2e7c990207f7564ea575a455e07, the /etc/issue banner
generation was ported back to the version from ptxdist-2018.09.0.

It turned out that there is a bug in ptxdist-2018.10.0 (missing ,g in
regular expression), so we go back to that version, with the missing ,g
added.

The fix should go into ptxdist-2018.11.0, so we can remove this local
override afterwards.

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>