platform-rpi: Enable early debug UART
This config enables the early debug for the Raspberry-Pi Firmware
bootloader.

This makes sure that the UART is available to barebox even if the
used board has additional hardware (like Bluetooth) connected to the
UART.

This command only works if the firmware is loaded from the SD-Card.

See for more information:
https://github.com/raspberrypi/firmware/wiki/USB-MSD-and-network-boot-debugging

Signed-off-by: Chris Fiege <chris@tinyhost.de>