platform-v7a: add patch for omap hwrng
The kernel does not use the omap hwrng as an entropy source, since the
quality is not set. The getrandom system call is now used by OpenSSH,
which results in a long wait of 2 Minutes on first boot because of the
ed25519 key generation.
This patch sets the quality within the driver, allowing the beaglebone
black to boot fast again.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>