v7a: barebox-common: add a common reference config for all barebox variants
This package serves multiple purposes:

* Make the barebox version and MD5 for all used barebox variants
  configurable with just one menu entry.
* Have a common barebox configuration with a baseline feature set, which
  can be used by all barebox-* variants as a reference config with the
  config diff functionality introduced in PTXdist 2019.01.0.
  The common config is currently based on ARCH_BCM283X, which seems to
  generate the smallest config diff for all the existing device-specific
  barebox configs (which will be adapted in the next commits).

The new barebox-common package only exists for the ability to call
"ptxdist menuconfig" on it in order to change the common config. It
isn't meant to build or install anything, so the compile, install and
targetinstall stages can be no-ops.

Also make sure that the menu entry for the common configuration is
listed before all other barebox variants by naming the kconfig file
appriopriately. Since the barebox-common config uses LZO image
compression, depend on host-lzop too.

Then configure it to use the version which is currently used for all
other variants in the v7a platform.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>