v7a: bareboxes: use shared settings from the barebox-common package
As implemented in the previous commit, the barebox version and MD5
can be reduced to one entry in the menu. Also do the same for the
host-lzop dependency, which will be needed for every barebox package
when using the common reference config. Be sure to depend on the new
barebox-common package so the barebox packages are rebuilt when the
version and MD5 variables change.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>