v7a: bareboxes: use config diffs with the common reference config
Make it easier to maintain a common set of features and also review the
device-specific changes between all the barebox variants by using the
config diff mechanism which has been available since PTXdist 2019.01.0.

No barebox config was harmed in the making of this patch, only their
difference to the common reference config was computed.

Also add a helper script to regenerate the config diffs after the common
config was changed.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>