datapartition: disable autoformat of non-MBR Linux partitions
The upcoming STM32MP157C-DK2 board will be the first DistroKit board
with a GPT based image and it has barebox as partition 3, unlike the
other boards where this is always the data partition.

This causes the data partition rule to brick the image by overwriting
barebox.

As a temporary measure, disable the data partition auto formatting
on GPT-images by explicitly checking for partition type 0x83 before
starting the service. Eventually we might want to match against
a UUID before autoformatting, or create the partition anew altogether.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>