MIPS: kernel: keep the config files stable
This is basically the same as the ptxdist upstream commit
c065184961ab1973c304ee31347205569b4f7536 ("kernel: add workaround to keep
kernelconfigs stable")

With linux-4.19 and gcc-8.x oldconfig will prompt for extra options that
are not available with e.g. gcc-7.x. This causes problems when a BSP is
built with different host compilers.
Add a workaround to optionally disable these extra options by forcing the
compiler check to fail.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>