v7a: tf-a: version bump v2.2 -> v2.3
TF-A v2.3 brings in support for multi-image STM32MP1 builds, which means
less boilerplate for us, because we can build images for multiple
STM32MP1 boards at once with the same rule. We want to do this in the
follow-up commit, so bump the version now.

With this update, all boot device drivers are disabled by default.
We can afford continuing to waste a few kilobytes in the first stage
bootloader and in the future, we might want to support boards that
don't boot from SDMMC anyway, so just (re-)enable all of them.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>