v7a: tf-a: add support for Linux Automation MC-1
The MC-1 can boot from SD-Card with the DK2 TF-A, but for use in
production, it's meant to boot from eMMC. Patch in the MC-1 device tree,
which enables the appropriate SDMMC controller.

RAM timings are those of the DK-2, which have worked so far, but should
be reconsidered prior to upstreaming.

[Note: eMMC boot doesn't yet work, presumably due to hardware issues.]

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>