v7a: kernel: configure for MC-1
Our current kernel config already supports one STM32MP board, but we
have some HW on the MC-1, that the DK-2 lacks, namely a KSZ9031 PHY,
an EEPROM in the SiP and PWM LEDs.

We also got a display controller, but in interest of keeping kernel
code size down, enabling DRM and associated drivers is left to a
ptxdist layer specific to the MC-1.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>