v7a: stm32mp: add Linux Automation MC-1 device tree
The device tree is based on the one in linux-stm32/stm32-next, which
will probably be merged for Linux v5.8-rc1. Instead of waiting that
long, we import it here with some stuff removed/changed, so it's usable
without the prerequisite patches.

We can drop them again, when we upgrade to v5.8-rc1.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>