v7a: barebox: add support for Linux Automation MC-1
This adds barebox support for the Linux Automation GmbH MC-1 board built
around the Octavo Systems OSD32MP157C-512M SiP.

The patches have been posted[1] upstream, but are not yet accepted.
It's however expected for them to be part of barebox v2020.06.0, at which
time we can drop the patchset again.

[1]: https://lists.infradead.org/pipermail/barebox/2020-April/041653.html

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>