v7a: images: stm32mp: remove cargo-culted partition-type in GPT config
partition-type is specific to DOS partitions. The counterpart for GPT
partitions would be partition-type-uuid, which already defaults to "L",
which is ok for our purposes.

So far, this didn't have an impact, but starting with genimage v13,
commit 013b22f26b73 ("Add support for Hybrid MBR partitions"), support
for a partition-type key in a GPT partition was introduced:

  Using this option with a GPT partition table will create a hybrid
  MBR partition  table with a maximum of 3 partition entries(this
  limit does not effect the maximum number of GPT partition entries
  in the same image).

The ARM TF-A can't cope with this new format and fails to boot barebox
with a "ERROR:   Partition ssbl not found" message. Drop partition-type
to prevent the ptxdist 2020.06.0 upgrade that includes host-genimage v13
from breaking the stm32mp157c-dk2 and lxa-mc1 images.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>