v7a: barebox: version bump v2020.04.0 → v2020.06.0
This allows us to drop the patch stack we had for Linux Automation MC-1
support. The series went into barebox with a fix that makes a warning
about the eMMC environment go away. In return we import a single patch
to address a breakage on the rpi3b platform. Being a bug fix, we should
be able to drop the patch along with v2020.07.0.

On the Kconfig front, CONFIG_OF_NET was dropped upstream, because it was
an unreferenced symbol. All Kconfig changes are the result of an
olddefconfig.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>