platform-v7a: kernel patch to fix PHY warnings
With the recent move to Linux v5.8, we get a new kernel warning
because it skew timings are now specified twice; once from use of
"rgmii-id" and one from the [rt]xc-skew-ps in the device tree.

This patch removes those timings since the board works fine
with the new default settings.

Patch has also been sent[1] to the respective mailing lists for upstream
inclusion.

[1]: https://lore.kernel.org/linux-arm-kernel/20200807150355.6116-1-h.assmann@pengutronix.de/

Signed-off-by: Holger Assmann <h.assmann@pengutronix.de>