v7a: barebox: version bump v2020.08.1 → v2020.09.0
The previous upgrade to v2020.08.1 gave us first stage boot support on
the sama5d2, but had slight multi-image breakage because the fresh
giantboard support on the sama5d27 messed up the EK1 environment
support: it set nv.boot.default=mmc1 in the built-in environment when
none should've been set. This is fixed in v2020.09.0, which is imported
here.

The config changes are those of olddefconfig.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>