v7a: barebox update 2020.09.0 -> 2021.02.0
One new patch is required to fix a regression for USB on rpi3. Also USB
on rpi3 requires CONFIG_USB_NOP_XCEIV=y now which is enabled in the rpi2
config now.

For the beaglebone two patches are added fixing a regression related to
a dts sync from upstream.

All three patches were already sent to the barebox mailing list.

Acked-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>