v7a: run: make the SSH forwarding port multiprocess-compatible
When multiple ./configs/platform-v7a/run scripts are run at the same
time, the first qemu process will open port 4445, and following runs
will fail. To prevent that, first look at the environment variable
QEMU_SSH_PORT to determine the port to be used (so that users can set
this in their shell to keep the port constant), or fall back to a port
between 1025 and 65025 based on the shell's PID, which should have a
good chance of choosing an unused, but stable port.

Also print the port forwarding before starting qemu, so we don't have to
bother about grepping the process list.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.pengutronix.de/20210426100714.10944-1-rhi@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>