v7a: barebox: version bump v2021.04.0 → v2021.05.0
Updating to v2021.05.0 lets us drop the single patch we have,
as it went upstream in revised form.

Kconfig changes are those produced by oldconfig. One notable
new option is CONFIG_DEFAULT_ENVIRONMENT_GENERIC_NEW_IKCONFIG,
which allows barebox to embed its configuration as /env/data/config
analogously to /proc/config.gz, which we already enable for Linux.
This is not enabled here, but might be something we want to use
eventually.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20210602111803.15832-1-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>