v7a: barebox: version bump v2021.05.0 → v2021.07.0
barebox v2021.07.0 now comes with support for first stage boot on the
SAMA5D3. The SAMA5D3 has a v7a CPU but lacks NEON, so it doesn't make
sense to enable for the v7a platform. Unfortunately, barebox master
is upset if you don't build both sama5d2 and sama5d3 at the same time.
Instead of working around this in the defconfig and enabling a board
we don't need and may forget to disable, import a small patch.

We should be able to drop this with v2021.08.0.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20210722133520.17946-1-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>