v7a: generate image for STM32MP7C-EV1
The STM32MP157C-EV1 Evaluation board is the full-feature demonstration
and development platform for STMicroelectronics Arm®-based dual
Cortex®-A7 32 bits and Cortex®-M4 32 bits MPUs in the STM32MP1 Series.

The small siblings DK1 and DK2 are already supported. Add now support
for the EV1 as well, so DistroKit can be used out-of-the-box with
all the official STM32MP1 development kits.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20210722133520.17946-4-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>