v7a: barebox: update v2021.11.0 → v2022.01.0
Freshest barebox release incorporates the OMAP patches we have in the
BSP. Bump up the release and drop these two patches in return.

The release also introduced CONFIG_STM32_BSEC_WRITE, which enables
blowing e-fuses on STM32MP1 SoCs, e.g. to persist an Ethernet address.
We support a couple of STM32MP1 boards, so it makes sense to enable
this option. All other Kconfig changes are those of olddefconfig.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>