v7a: tf-a: update to v2.7
We need at least v2.6 for FIP support, but there's a freshly released
v2.7, so lets use that. This unfortunately means we need to carry a
patch for the MC-1 because upstream broke it.  To ensure the legacy
image works for now, we need to set a new build option, but that will
be removed soon enough with the switch to FIP.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20220725120445.1295669-2-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>