v7a: barebox-stm32mp: start collecting device tree blobs
FIP images are passed along a device tree via r2 (called hardware
config). As we want barebox to be booted with the barebox DT to
decouple kernel and bootloader updates, install the relevant barebox
DTs, but with a barebox- prefix as they would conflict with kernel
device tree blobs otherwise.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20220725120445.1295669-4-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>