v7a: tf-a: start collecting FIP image components
For building FIP images, we need a firmware config describing e.g.
where to place components, a device tree for the bl32, a secure
monitor (SP_min, bl32.bin), as well as non-trusted firmware and
hardware configuration. The latter two will be provided by barebox,
so install here the ones provided by TF-A in addition to the TF-A
binaries itself (which will load the FIP).

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20220725120445.1295669-5-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>