host-genimage: version bump 15 -> 16
Quoting Ahmad Fatoum's original patch description:

"
genimage master has support for calling fiptool, which makes it
straight-forward to generate eMMC boot partition images with the
same genimage.config generating the image for the user area.

TF-A also has a fip target that can call fiptool for us, but is only
works when generating a single bootloader image, but we generate
multiple images, so we can't reuse the current tf-a rule as is.

One more option would be to write a custom image rule, but as genimage
master has support, let's just backport the patch. The autogen.sh is
necessary, because we touch Makefile.am and we want autogen to rerun to
account for changes between our host autotools and the one used to
build the shipped files.
"

Instead of patching genimage, let's bump the version.

This patch and the one that pulls in the host-genimage.make rule can be
removed from DistroKit once genimage-16 is in ptxdist.

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>