v8a: rockchip: add rule for rockchip firmware
There are prebuilt binaries by Rockchip for SDRAM configuration of the
board we use. Add a rule to download them off Github.

This rule also downloads BL31 (TF-A) and BL32 (OP-TEE) binaries, which
we will eventually compile from source. However, as of now, Rockchip did
not release the source code and we have to use use the prebuilt
binaries. Any board specific setup we need will happen within Barebox.

Signed-off-by: Johannes Zink <j.zink@pengutronix.de>
Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>
Link: https://lore.pengutronix.de/20220905163615.2807729-15-m.tretter@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>