barebox: deploy new barebox-raspberry-pi.img
The generic barebox-dt-2nd.img mimics Linux and allows booting barebox
with an external device tree as if it were a kernel. We employ this for
using the same image for Rpi 2 and 3 by placing barebox device trees
into the FAT boot partition and having the VideoCore pass the correct
device tree to barebox.

While this works, it means the VideoCore won't patch the kernel device
tree, but many users depend on the VideoCore to apply overlays for
kernel consumption. While this could be done in barebox instead, it
doesn't have the familiar config.txt interface and it doesn't cover some
of the more obscure magic the videocore does: e.g. when applying an
overlay configuring the USBOTG, the VideCore will reconfigure clocks as
a side effect of applying the overlay.

For this reason, barebox now also has a barebox-raspberry-pi.img that
behaves the same as barebox-dt-2nd.img with the difference that it has
the multiple device trees built-in. It's already being generated since
updating to the new release, so let's make use of this.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20220930162159.814389-6-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>