v7a: rpi4: place DT into /boot for VideoCore consumption
This is a somewhat ugly hack. The VideoCore will pass this device tree
to barebox, which doesn't care much for it. As a side effect of the
device tree passing, the VideoCore parses the device tree and will
detect that the mini-uart is enabled and take care not to scale clock
rate on the VideoCore, which the mini-uart is sourced from as that would
upset the baudrate.

There may be some config.txt entry we can use to avoid this, but for
now, just place the device tree into the boot partition.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20220930162159.814389-10-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>