doc: v8a: ROCK 3A: add documentation on writing barebox-only images to SD-Card and NOR Flash
It may be useful to sometimes only write a barebox image to SD-Card,
e.g. for chainloading, TFTP-Boot or if Android Fastboot is used for
flashing the actual full image to a storage medium. Add documentation
on which image to use.

Also add a short section about how to write barebox to the fallback
NOR SPI Flash on the ROCK 3A board.

Signed-off-by: Johannes Zink <j.zink@pengutronix.de>
Link: https://lore.pengutronix.de/20230324182952.501117-1-j.zink@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>