doc: replace "copy onto" by "write to" to prevent confusion
"Copy" could be misunderstood as "use cp (on the file system level)",
which will not lead to the intended results, i.e. writing to the device
directly (e.g. using dd).

Furthermore, using "write onto" [1] instead of "write to" [2] feels
unnatural here – as if writing something on top of another. Use "to"
instead.

[1]: about 460.000 Google results
[2]: about 146.000.000 Google results

Signed-off-by: Roland Hieber <rhi@pengutronix.de>