ptxconfig: enable libatomic
Since openssl 3.x, they link agains libatomic unconditionally, and we build
with -Wl,--as-needed so it gets dropped on platforms that don't need it.

This triggers a reason warning on platform-rpi1 and platform-mips.

Enable libatomic locally and whitelist the reason check. This adds a
small library, but shouldn't harm otherwhise.

Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>