v7a: at91bootstrap: configure v4.0.6-rc1 for Wifx L1
We will use AT91Bootstrap as first stage bootloader for the Wifx L1 for
now, so let's add a suitable config file. We import v4.0.6-rc1 as that's
the first tag that contains commit 41cd6c4 ("crt0_gnu.S: always pass
along r4 to next stage"), which is required for later boot firmware to
determine what the initial boot medium was.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20230417131141.908678-3-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>