v7a: kernel: extend support to SAMA5D4-based Wifx L1
Let's import the barebox device tree into the BSP for use by Linux.
This allows us to update it separately in the future. The kernel was so
far configured for SAMA5D2 only, so enable whatever else is needed for
the SAMA5D4. While we boot from SD for now, main boot medium is meant to
be NAND. We thus enable the NAND driver and UBIFS. Driver support is
there, but we have yet to integrate boot from NAND into the BSP.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20230417131141.908678-5-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>