doc: use layered documentation instead of overwriting doc/guru.rsto
Since PTXdist 2023.01.0, all files named index-layer-*.rst are included
in the main toctree by upstream PTXdist, so there is no need anymore to
fork the upstream guru.rst. See the docs for more info.

We have to remove the :numbered: argument because the main toctree in
the upstream guru.rst is already numbered, otherwise this will generate
a warning.

Link: https://www.ptxdist.org/doc/daily_work_section.html#integrate-project-specific-documentation-into-the-manual
Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.barebox.org/20230510133915.1724757-1-rhi@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>