v8a: barebox: rock3a: drop no longer needed SD-Card speed quirk
SD-Card signal integrity on the Rock-3A's board doesn't allow for
reliable sd-uhs-sdr104 operation. For this reason we had a device tree
fixup that removed the property. This has been fixed upstream in the
meantime and sd-uhs-sdr104 was replaced with the slower sd-uhs-sdr50,
which works fine for us, even with a usbsdmux in-between.

Thus drop this fixup again.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>