v8a: barebox: version bump v2023.01.0 → v2023.05.0
As preparation for combining barebox-rockchip and barebox-imx8m, let's
update both to v2023.05.0, which is the second release that can target
both with the same build.

This barebox update also pulls in the kernel DT update that moved
around the DT aliases for the Radxa boards. For this reason, we need
to touch up the environment as well.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>