v8a: barebox: configure upstream barebox rule for both imx8m and rockchip
So far, we duplicated the barebox rules, so we can customize them for
each of imx8m and rockchip platforms. With v2023.04.0, we can configure
the same barebox build to generate binaries for both at the cost of
compiling in some unused drivers (but not duplicating DT). Let's do
that.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>