v8a: configure MTD SPI-NOR sector sizes consistently
It must be guaranteed that the option MTD_SPI_NOR_USE_4K_SECTORS is set
consistently between barebox and kernel. If this is not the case bad things
can happen (e.g. inconsistencies in barebox state).

The platforms we have use MMC by default, but let's take no chances and
do (one of the) right thing(s) by aligning barebox with the kernel,
which is already built with CONFIG_MTD_SPI_NOR_USE_4K_SECTORS=y.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>