v8a: barebox: add missing dependency on TF_A for i.MX
Unlike firmware-rockchip, firmware-imx doesn't ship BL31 (TF-A). Instead
we configure the upstream TF-A rule to build images for i.MX8M Quad,
Mini and Plus.

With the switch to the upstream barebox rule, this dependency on TF-A
got lost breaking less-parallel fresh builds, so let's add it back.

Preferably, TF_A would be selected directly from the barebox rule's
Kconfig. There's a BAREBOX_DEPENDENCIES option for that, but I don't
know if it's possible to extend barebox.in without having to duplicate
it fully, so adding it to firmware-imx should suffice for now.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>