v8a: barebox: enable loading kernel, dtb and initrd via fastboot
This allows to load kernel, dtb and initrd as sparse images to
RAM-files.

While at it, also enable the acm gadget, which exposes the barebox shell
as an USB gadget and autostart the usb gadgets.

Also expose mmc{0,1,2} as fastboot partitions and add init script to
force barebox to detect these devices and have them available for
fastboot.

Signed-off-by: Johannes Zink <j.zink@pengutronix.de>