platform-v8a: cleanup barebox firmware handling
BAREBOX_INJECT_FILES is explicitly designed in such a way that all
packages can add to it in their regular rules, so move the declarations
accordingly. And only add to BAREBOX_INJECT_FILES if the package is
acutally enabled.

imx/imx-usb-loader is already part of BAREBOX_PROGS_HOST, so no need to
add it.

Move rk-usb-loader to rules/barebox.rockchip.make (it's explicitly
loaded after rules/barebox.make) for now. This can be removed when this
is added upstream.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>
Link: https://lore.barebox.org/20230620132931.2179983-2-m.olbrich@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>