x86_64: run: add 9p tcp based rootfs support
Since there is no diod server ptxdist host tool available yet, this
script starts the version from debian.

It is also to be found how to best start the server.
For now this simple script works as proof of concept including
ugly hacks to run and stop diod around the qemu exec.

Signed-off-by: Michael Grzeschik <m.grzeschik@pengutronix.de>
Link: https://lore.pengutronix.de/distrokit/20230621224609.932551-1-m.grzeschik@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>