platforms: revert to default systemd loglevel
By decreasing the systemd loglevel to "warning", the journal no longer
contains useful lines like which units were started and stopped and
when, which is quite sparse information for debugging. Remove the
systemd.log_level from the kernel command line so systemd defaults to
loglevel "info" (same level as loglevel=5 for the kernel).

Keep the systemd.show_status setting on "auto" so that systemd only
produces actual console output when the boot takes a significant amount
of time, or produces error messages; so this change only changes the
verbosity of the journal, not the console, on normal boots.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.barebox.org/20230623124755.2292833-2-rhi@pengutronix.de
[remove meanwile eliminated special environment for rock3a, see 8ed4f77e790a24bd8c0d4fb15f95731eb1490be6 for reference]
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>