v7a: rpi2: Restore finding rpi-firmware when used as base layer
When DistroKit was used as a base layer, the firmware was not in
'$(PTXDIST_WORKSPACE)/rpi-firmware' but in
'$(PTXDIST_WORKSPACE)/base/rpi-firmware' or
'$(PTXDIST_WORKSPACE)/base/base/rpi-firmware' … so we just need to look
for it in all layers.  Otherwise no firmware is put to the FAT partition
of the SD card image and the device does not boot at all.

Fixes: 4dbcbb498769 ("v7a: rpi2: move rpi-firmware directory to top-level workspace")
Signed-off-by: Alexander Dahl <ada@thorsis.com>
Reviewed-by: Michael Olbrich <m.olbrich@pengutronix.de>
Link: https://lore.barebox.org/20230630120847.3271567-1-ada@thorsis.com
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>