v8a: imx8m: genimage: switch from MBR to GPT images
MBR is stuff of the past. Switch over to GPT. This is easier on i.MX8M
than on i.MX6, because the BootROM starts execution at an >= 32K offset,
so GPT can be located at its usual location. We should eventually move
barebox into a proper GPT partition, but for now, just keep the old
partition numbering intact.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Reviewed-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.barebox.org/20230728164933.342061-6-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>