v8a: imx8m: include i.MX8MN DDR4 and LPDDR4 EVK DT in rootfs
Now that we are building barebox for the board, enable installation of
the kernel device trees, so we can boot a kernel with it. The necessary
kernel options are already enabled. As DDR4 and LPDDR4 differ in the
PMIC used, we enable two different device trees. barebox will choose the
correct one by comparing their DT compatible against its own.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Reviewed-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.barebox.org/20230728164933.342061-7-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>