silence QA checks regarding missing kernel features for RAUC
Since we can only select RAUC in the ptxconfig for all platforms, PTX
internal QA tooling warns that necessary features for RAUC are missing
in the kernelconfigs on non-v7a platforms.

The issue of overriding the tests on the platform level seems a bit
harder to fix in the QA tool, so until that can be done, please the QA
checker by enabling the necessary features in all remaining kernels to
make the PTX Jenkins tests green again.

If the QA tool is fixed, this commit can be reverted.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.barebox.org/20230825110310.2948272-4-rhi@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>