image-rauc: use ext4 rootfs instead of tar.gz
On one hand, the squashfs inside the RAUC bundle already uses
compression, so with a tar.gz inside a squashfs the content is
compressed twice. Also using ext4 makes it possible to use adaptive
updates in the future.

(Yes, the image generated by the image-root-ext recipe is really called
"root.ext2", even if it's an ext4 image.)

Link: https://rauc.readthedocs.io/en/latest/advanced.html#adaptive-updates
Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.barebox.org/20230825110310.2948272-7-rhi@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>