v7a: barebox: enable loading kernel, dtb and initrd via fastboot
This change is similar to:

2732f93bcf4ee011cecca3518aec7e523c311d7a
 "v8a: barebox: enable loading kernel, dtb and initrd via fastboot"

but for some of the v7a platforms. Due to the absence of a common environment
it only can be added for the i.MX6 based systems for now.

This change prepares all v7a platforms to load kernel, dtb and initrd as
sparse images to RAM-files for rapid development.
It exposes mmc{1,2,3} as 'fastboot' partitions and add init script to
force barebox to detect these devices and have them available for
'fastboot'.

It autostarts the USB gadget and network capability to use 'fastboot' out of
the box.

Signed-off-by: Juergen Borleis <jbe@pengutronix.de>
Link: https://lore.barebox.org/20230825111056.32801-1-jbe@pengutronix.de

Adapted bsp.ref for v7a, similar to v8a.

Inspired-by: Roland Hieber <rhi@pengutronix.de>
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>