qemu-run: don't evaluate selected_platformconfig
Even when a selected_platformconfig symlink exists, ptxdist can still be
used with the --platformconfig=... option to override the platformconfig
during the build. In this case, running the qemu script will search in
the wrong build folder for the images.

Don't use the selected_platformconfig link to find out the platform,
instead determine it from the path of the qemu script itself.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.pengutronix.de/distrokit/20231010091028.3444582-1-rhi@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>