v7a: barebox: pass root=mmcblkXpN instead of root=PARTUUID=… to kernel
Kernels prior to v5.10-rc1 could not ensure deterministic ordering for
MMC devices. For this reason, the default 'append-root' behaviour of
barebox for MMC block devices is to append a 'root=PARTUUID=XYZ' option
to the kernel command line. The PARTUUID is derived from the partition
table and is expected to be unique system-wide. This is however
susceptible to breakage if the exact same image is written to different
boot media in the system, e.g. SD and eMMC, so the PARTUUID will no
longer uniquely refer to a single partition.

Newer kernel versions honor the numbering used for 'mmc' aliases,
which makes 'root=mmcblkXpN' deterministic as long as barebox and
kernel aliases are identical. Since we're using kernel 6.4, enable this
option in barebox.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.pengutronix.de/20231020090519.3320460-3-rhi@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>